fix build

+++ 2.Firmware/PlatformIO/Peak-ESP32-fw/src/HAL/HAL_Def.h
@@ -22,18 +22,18 @@ typedef struct
 /* IMU */
 typedef struct
 {
-    float ax;
-    float ay;
-    float az;
-    float gx;
-    float gy;
-    float gz;
-    float mx;
-    float my;
-    float mz;
-    float roll;
-    float yaw;
-    float pitch;
+    int16_t ax;
+    int16_t ay;
+    int16_t az;
+    int16_t gx;
+    int16_t gy;
+    int16_t gz;
+    int16_t mx;
+    int16_t my;
+    int16_t mz;
+    int16_t roll;
+    int16_t yaw;
+    int16_t pitch;
 } IMU_Info_t;
 
 /* Power */
